--- a/runtime/compiler/codegen/J9RecognizedMethodsEnum.hpp
+++ b/runtime/compiler/codegen/J9RecognizedMethodsEnum.hpp
@@ -461,6 +461,7 @@
    jdk_internal_misc_Unsafe_copyMemory0,
    jdk_internal_loader_NativeLibraries_load,
    jdk_internal_util_ArraysSupport_vectorizedMismatch,
+   jdk_internal_util_ArraysSupport_vectorizedHashCode,
    jdk_internal_util_Preconditions_checkIndex,
 
    FirstVectorMethod,