`ArraysSupport.vectorizedHashCode` [1] is an intrinsic candidate introduced in JDK21. The hash function used is the same as the one used for hashing Strings, so this change leverages that by recognizing `ArraysSupport.vectorizedHashCode` and using the preexisting accelerated implementation of `String.hashCode`.

[1] https://github.com/ibmruntimes/openj9-openjdk-jdk21/blob/openj9/src/java.base/share/classes/jdk/internal/util/ArraysSupport.java#L200