@Spencer-Comin 2024-05-16
Here's a couple instruction selection logs:
### Integer array
```
------------------------------
 n9n      (  0)  treetop                                                                              [     0x3ff48f19910] bci=[-1,6,40] rc=0 vc=106 vn=- li=2 udi=- nc=1
 n8n      (  1)    icall  jdk/internal/util/ArraysSupport.vectorizedHashCode(Ljava/lang/Object;IIII)I[#408  static Method] [flags 0x500 0x0 ] (in GPR_0046) ()  [     0x3ff48f198c0] bci=[-1,6,40] rc=1 vc=106 vn=- li=2 udi=- nc=5 flg
 n3n      (  0)      aload  <parm 0 Ljava/lang/Object;>[#399  Parm] [flags 0xc0000107 0x0 ] (in &GPR_0032)  [     0x3ff48f19730] bci=[-1,0,40] rc=0 vc=106 vn=- li=2 udi=- nc=0
 n4n      (  0)      iload  <parm 1 I>[#400  Parm] [flags 0xc0000103 0x0 ] (in GPR_0036) (cannotOverflow )  [     0x3ff48f19780] bci=[-1,1,40] rc=0 vc=106 vn=- li=2 udi=- nc=0 flg=0x1000
 n5n      (  0)      iload  <parm 2 I>[#401  Parm] [flags 0xc0000103 0x0 ] (in GPR_0039) (cannotOverflow )  [     0x3ff48f197d0] bci=[-1,2,40] rc=0 vc=106 vn=- li=2 udi=- nc=0 flg=0x1000
 n6n      (  0)      iload  <parm 3 I>[#402  Parm] [flags 0xc0000103 0x0 ] (in GPR_0046) (cannotOverflow )  [     0x3ff48f19820] bci=[-1,3,40] rc=0 vc=106 vn=- li=2 udi=- nc=0 flg=0x1000
 n7n      (  0)      iconst 10 (X!=0 X>=0 )                                                           [     0x3ff48f19870] bci=[-1,4,40] rc=0 vc=106 vn=- li=2 udi=- nc=0 flg=0x104
------------------------------

LG      &GPR_0032, Parm[<parm 0 Ljava/lang/Object;>] ?+0(GPR5)
L       GPR_0036, Parm[<parm 1 I>] ?+8(GPR5)
L       GPR_0039, Parm[<parm 2 I>] ?+16(GPR5)
LGFR    GPR_0036,GPR_0036
LGFR    GPR_0039,GPR_0039
L       GPR_0046, Parm[<parm 3 I>] ?+24(GPR5)
Label L0032:     # (Start of internal control flow)
SLLG    GPR_0036,GPR_0036,2
SLLG    GPR_0039,GPR_0039,2
LA      GPR_0043,#410 0(GPR_0039,GPR_0036)
CLGIJ   GPR_0039,Label L0036,16,BM(mask=0x4),
Label L0033:
SLGFI   GPR_0043,12
LARL    GPR_0047, &<LiteralPool Base Address>     ; LoadLitPool
VL      VRF_0040,#411 =X(0) 0(GPR_0047),0
VGBM    VRF_0041,0x0
VLVGF   VRF_0041,GPR_0046,#412 3
Label L0034:
VL      VRF_0042,#413 8(GPR_0036,&GPR_0032),0
VMALF   VRF_0041,VRF_0041,VRF_0040,VRF_0042
LA      GPR_0036,#414 16(GPR_0036)
CLGRJ   GPR_0036,GPR_0043,Label L0034,MASK5(mask=0x4),
Label L0035:
LARL    GPR_0048, &<LiteralPool Base Address>     ; LoadLitPool
VL      VRF_0040,#411 =X(0) 0(GPR_0048),0
VMLF    VRF_0041,VRF_0041,VRF_0040
VGBM    VRF_0040,0x0
VSUMQF  VRF_0040,VRF_0041,VRF_0040
LA      GPR_0043,#415 12(GPR_0043)
VLGVF   GPR_0046,VRF_0040,#416 3
Label L0036:     # (End of internal control flow)
CLGRJ   GPR_0036,GPR_0043,Label L0038,MASK11(mask=0xa),
Label L0037:     # (Start of internal control flow)
SLLK    GPR_0039,GPR_0046,5
SGR     GPR_0039,GPR_0046
L       GPR_0046,#417 8(GPR_0036,&GPR_0032)
LA      GPR_0036,#418 4(GPR_0036)
AGR     GPR_0046,GPR_0039
CLGRJ   GPR_0036,GPR_0043,Label L0037,MASK5(mask=0x4),
assocreg
Label L0038:     # (End of internal control flow)
```



### Byte array
```
------------------------------
 n9n      (  0)  treetop                                                                              [     0x3ff48f19910] bci=[-1,6,32] rc=0 vc=95 vn=- li=2 udi=- nc=1
 n8n      (  1)    icall  jdk/internal/util/ArraysSupport.vectorizedHashCode(Ljava/lang/Object;IIII)I[#408  static Method] [flags 0x500 0x0 ] (in GPR_0025) ()  [     0x3ff48f198c0] bci=[-1,6,32] rc=1 vc=95 vn=- li=2 udi=- nc=5 flg=
 n3n      (  0)      ==>aRegLoad (in &GPR_0016)
 n4n      (  0)      ==>iRegLoad (in GPR_0017) (cannotOverflow )
 n5n      (  0)      ==>iRegLoad (in GPR_0018) (cannotOverflow )
 n6n      (  0)      iload  <parm 3 I>[#402  Parm] [flags 0xc0000103 0x0 ] (in GPR_0025) (cannotOverflow )  [     0x3ff48f19820] bci=[-1,3,32] rc=0 vc=95 vn=- li=2 udi=- nc=0 flg=0x1000
 n7n      (  0)      iconst 8 (X!=0 X>=0 )                                                            [     0x3ff48f19870] bci=[-1,4,32] rc=0 vc=95 vn=- li=2 udi=- nc=0 flg=0x104
------------------------------

LGFR    GPR_0017,GPR_0017
LGFR    GPR_0018,GPR_0018
L       GPR_0025, Parm[<parm 3 I>] ?+24(GPR5)
Label L0016:     # (Start of internal control flow)
LA      GPR_0022,#412 0(GPR_0018,GPR_0017)
CLGIJ   GPR_0018,Label L0020,4,BM(mask=0x4),
Label L0017:
SLGFI   GPR_0022,3
LARL    GPR_0026, &<LiteralPool Base Address>     ; LoadLitPool
VL      VRF_0019,#413 =X(0) 0(GPR_0026),0
VGBM    VRF_0020,0x0
VLVGF   VRF_0020,GPR_0025,#414 3
Label L0018:
VLLEZF  VRF_0021,#415 8(GPR_0017,&GPR_0016)
VUPHB   VRF_0021,VRF_0021
VUPLHW  VRF_0021,VRF_0021
VMALF   VRF_0020,VRF_0020,VRF_0019,VRF_0021
LA      GPR_0017,#416 4(GPR_0017)
CLGRJ   GPR_0017,GPR_0022,Label L0018,MASK5(mask=0x4),
Label L0019:
LARL    GPR_0027, &<LiteralPool Base Address>     ; LoadLitPool
VL      VRF_0019,#413 =X(0) 0(GPR_0027),0
VMLF    VRF_0020,VRF_0020,VRF_0019
VGBM    VRF_0019,0x0
VSUMQF  VRF_0019,VRF_0020,VRF_0019
LA      GPR_0022,#417 3(GPR_0022)
VLGVF   GPR_0025,VRF_0019,#418 3
Label L0020:     # (End of internal control flow)
CLGRJ   GPR_0017,GPR_0022,Label L0022,MASK11(mask=0xa),
Label L0021:     # (Start of internal control flow)
SLLK    GPR_0018,GPR_0025,5
SGR     GPR_0018,GPR_0025
LB      GPR_0025,#419 8(GPR_0017,&GPR_0016)
LA      GPR_0017,#420 1(GPR_0017)
AGR     GPR_0025,GPR_0018
CLGRJ   GPR_0017,GPR_0022,Label L0021,MASK5(mask=0x4),
assocreg
Label L0022:     # (End of internal control flow)
```