@r30shah 2024-05-24
jenkins test sanity zlinux jdk17,jdk21