@r30shah 2024-05-24
@Spencer-Comin  Given that this is accelerating the Arrays.hashcode implementation in JDK21 and onwards - I think it would be good idea to evaluate the performance through a benchmark (JMH based / bumblebench) to see how much speed up we got