@Spencer-Comin 2024-05-29
Here's some benchmarking results for `Arrays.hashcode` on  10,000,000 element arrays:

| Element Type | Baseline (ops/s) | Accelerated (ops/s) | Acceleration |
| ------------ | ---------------- | ------------------- | ------------ |
| `byte`       | 121.008          | 416.066             | 343.8%       |
| `char`       | 119.725          | 402.715             | 336.4%       |
| `int`        | 125.050          | 406.709             | 325.2%       |
| `short`      | 107.173          | 400.004             | 373.2%       |

We are getting a 3-4x improvement, which makes sense given that we go from consuming 1 element at a time in the scalar path to 4 at a time in the vector path.